ESP32 CAM ESP NOW
It needs to esp32 boards.
First one master common ESP32, with button connected to GPIO 21 and Ground. Which sends Command to Take a Picture.
Second one Slave ESP32 CAM, with SD card connected to it. whenever It recieves signal it will take a picture and save it in the folder photos, with prefix photo_x.